requirements.txt
python-telegram-bot==21.6
pdfplumber==0.11.4
python-docx==1.1.2
openpyxl==3.1.5
pytesseract==0.3.10
Pillow==10.4.0
python-dotenv==1.0.1
